requirements.txt
asgiref==3.8.1
attrs==24.2.0
blis==0.7.11
catalogue==1.0.2
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
cymem==2.0.8
Django==4.2.16
djangorestframework==3.15.2
docx2txt==0.8
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz#sha256=06c80936324012d1223291d2af41a5229e746dc2dee8fe31a532666ee3d18aaa
idna==3.8
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
murmurhash==1.0.10
nltk==3.9.1
numpy==1.23.5
pandas==2.2.2
pdfminer.six==20240706
plac==1.1.3
preshed==3.0.9
pycparser==2.22
pycryptodome==3.20.0
pyresparser==1.0.6
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
sortedcontainers==2.4.0
spacy==2.3.5
sqlparse==0.5.1
srsly==1.0.7
thinc==7.4.6
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wasabi==0.10.1
